+++ stories/table.stories.js
@@ -149,6 +149,13 @@ export const title = () => (
   <Table data={provinces} title='Cases by province' />
 )
 
+export const initialSortBy = () => (
+  <Table
+    data={provinces}
+    sortBy={{ id: 'new_cases', desc: true }}
+  />
+)
+
 export const tableProps = () => (
   <Table
     data={provinces}