+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `<Table />` - initial `sortBy` property for specifying one (Object) or multiple (Array of Objects) column sorting.
+
 ### Changed
 - `<Table />` hide toolbar when no title and no data.
 - `<Table />` Edit columns list with a maxHeight for in-list scroll.